Refactoring: move ChangeSet to its module
